Playground: Combobox
Everything is coming along nicely. In this demo, most of the UI works,
except the ToC and the page title. clean_page.php generates the page as
requested without any cruft - also returning a 404 if no folder exists.
next & previous works, as well as zoom and single / double page layouts.
left is always an even page, right is always an odd page (just like the
real world)  note: This is webkit only at the moment. I'm afraid to even
try it in firefox… but that's mostly because I haven't added -moz tags
for the css3 used.  if all goes well, this could be implemented in the
actual component shortly.