Minor tweaks
fixed up the new model a bit, still just playing around with it though.
Looked a bit how I want to generate the images, probably just gonna go
php + gd / or have hooks to imagemagick if supported. Never worked with
it before though (Only GD). Also, probably stick with JPG/PNG encoding,
base64 does not seem to be the way to go (not that I ever thought it
was, just poking at the idea)